+++ include/daScript/simulate/aot.h
@@ -1389,7 +1389,7 @@ namespace das {
         static __forceinline void clear ( Context * __context__, TTable<TKey,TVal> & tab ) {
             if ( tab.data ) {
                 if ( !tab.lock ) {
-                    uint32_t oldSize = tab.capacity*(sizeof(TKey)+sizeof(TVal)+sizeof(uint32_t));
+                    uint32_t oldSize = tab.capacity*(sizeof(TKey)+sizeof(TVal)+sizeof(uint64_t));
                     __context__->heap->free(tab.data, oldSize);
                 } else {
                     __context__->throw_error("can't delete locked table");
