+++ include/daScript/simulate/heap.h
@@ -581,7 +581,7 @@ namespace das {
             return size;
         }
         char * allocateCachedName ( const string & name );
-        das_hash_map<uint32_t,TypeInfo *>    lookup;
+        das_hash_map<uint64_t,TypeInfo *>    lookup;
         das_hash_map<string, char *>         stringLookup;
         uint64_t                             stringBytes = 0;
     };
